Split out the download of the tool from the clone This makes testing easier and allow a means to override the exe location Also fixed some typos in function file names
