+++ Extensions/WikiPDFExport/WikiPDFExportTask/src/WikiPDFExportTask.ts
@@ -2,7 +2,8 @@ import tl = require("azure-pipelines-task-lib");
 import * as fs from "fs";
 
 import {
-    ExportRun
+    ExportRun,
+    GetExePath
     } from "./ExportFunctions";
 
 var repo = tl.getInput("repo");
@@ -16,7 +17,8 @@ var useAgentToken = tl.getBoolInput("useAgentToken");
 var branch = tl.getInput("branch");
 var injectExtraHeader = tl.getBoolInput("injectExtraHeader");
 var cloneRepo = tl.getBoolInput("cloneRepo");
-var exeFolder = tl.getVariable("Agent.TempDirectory");
+var overrideExePath = tl.getInput("overrideExePath");
+var workingFolder = tl.getVariable("Agent.TempDirectory");
 
 console.log(`Variable: Repo [${repo}]`);
 console.log(`Variable: Use Agent Token [${useAgentToken}]`);
@@ -27,19 +29,26 @@ console.log(`Variable: SingleFile [${singleFile}]`);
 console.log(`Variable: OutputFile [${outputFile}]`);
 console.log(`Variable: Branch [${branch}]`);
 console.log(`Variable: InjectExtraHeader [${injectExtraHeader}]`);
-console.log(`Variable: Exe Download Folder [${exeFolder}]`);
+console.log(`Variable: OverrideExePath [${overrideExePath}]`);
 
-ExportRun(
-    exeFolder,
-    cloneRepo,
-    localpath,
-    singleFile,
-    outputFile,
-    extraParams,
-    useAgentToken,
-    repo,
-    user,
-    password,
-    injectExtraHeader,
-    branch
-);
+GetExePath(
+    overrideExePath,
+    workingFolder
+).then((exePath) => {
+    if (exePath.length > 0) {
+        ExportRun(
+            exePath,
+            cloneRepo,
+            localpath,
+            singleFile,
+            outputFile,
+            extraParams,
+            useAgentToken,
+            repo,
+            user,
+            password,
+            injectExtraHeader,
+            branch
+        );
+    }
+});
