+++ Extensions/WikiPDFExport/WikiPDFExportTask/test/geturl.test.ts
@@ -1,7 +1,7 @@
 import { expect } from "chai";
 // if you used the '@types/mocha' method to install mocha type definitions, uncomment the following line
 import "mocha";
-import { GetTrimmedUrl, GetProtocol } from "../src/GitWikiFuntions";
+import { GetTrimmedUrl, GetProtocol } from "../src/GitWikiFunctions";
 import { logInfo } from "../src/agentSpecific";
 import { existsSync, fstat } from "fs";
 import { pathToFileURL } from "url";
