+++ Extensions/WikiPDFExport/WikiPDFExportTask/test/testrunnerToCloneFromGitHub.ts
@@ -26,10 +26,10 @@ const singleFile = "";
 const extraParams = "";
 const useAgentToken = false;
 const cloneRepo = true;
-const exeFolder = "C:\\tmp";
+const exePath = "..\\testdata\\azuredevops-export-wiki.exe";
 
 ExportRun(
-    exeFolder,
+    exePath,
     cloneRepo,
     localpath,
     singleFile,
