+++ Extensions/WikiPDFExport/WikiPDFExportTask/testdata/dummy.exe.txt
@@ -1,0 +1,1 @@
+This is a text file to act as a place holder to check the exe file discovery logic
